Moves integration tests to unit tests. Merges conftest, removes nueron mock. Removes integration test from tox.ini. Adapst guthub actions to the changes.
