+++ .github/workflows/coverage.yml
@@ -39,13 +39,3 @@ jobs:
       uses: coverallsapp/github-action@v2
       with:
         flag-name: unit
-
-    - name: Run tox
-      run: |
-        rm -rf .coverage
-        tox -e integration
-
-    - name: Coveralls
-      uses: coverallsapp/github-action@v2
-      with:
-        flag-name: integration
