+++ tests/unit/test_stimulus2.py
