+++ tox.ini
@@ -2,7 +2,7 @@
 name = neurodamus
 
 [tox]
-envlist = flake8, format, unit, baseline, integration, integration-e2e, scientific
+envlist = flake8, format, unit, baseline, integration-e2e, scientific
 
 [testenv]
 deps =
@@ -28,13 +28,6 @@ commands =
     tests/simulations/v5_sonata/output* \
     sim_conf"
 
-[testenv:unit]
-deps =
-    {[testenv]deps}
-    pytest-cov
-commands =
-    pytest --cov=neurodamus --cov-report=term-missing tests/unit
-
 [testenv:baseline]
 deps =
     {[testenv]deps}
@@ -42,7 +35,7 @@ deps =
 commands =
     pytest --cov=neurodamus --cov-report=term-missing tests/test_import.py
 
-[testenv:integration]
+[testenv:unit]
 deps =
     {[testenv]deps}
     pytest-cov
@@ -55,7 +48,7 @@ allowlist_externals =
     sh
 commands =
     {toxinidir}/ci/build_ndcore.sh {toxinidir}/neurodamus/data {toxworkdir}
-    sh -c ". {toxworkdir}/.envfile && pytest -x --cov=neurodamus --cov-report=term-missing --forked tests/integration"
+    sh -c ". {toxworkdir}/.envfile && pytest -x --cov=neurodamus --cov-report=term-missing --forked tests/unit"
 
 [testenv:integration-e2e]
 # Please you need to spack load neurodamus-models model=neocortex +coreneuron or use an equivalent spack virtual env
